+++ src/placement/constraint_updater.h
@@ -32,7 +32,7 @@ class ConstraintUpdater
                                Eigen::Vector2i lastAnchorPosition,
                                Eigen::Vector2i lastLabelPosition,
                                Eigen::Vector2i lastLabelSize);
-  void drawRegionsForAnchors(std::vector<Eigen::Vector2i> anchorPositions,
+  void drawRegionsForAnchors(std::vector<Eigen::Vector2f> anchorPositions,
                              Eigen::Vector2i labelSize);
 
   void clear();
